added pool2 hooks
