+++ src/components/pages/Pool2/Pool2Page.tsx
@@ -1,3 +1,4 @@
+import React, { useState } from "react";
 import {
   useInterval,
   useDisclosure,
@@ -6,20 +7,30 @@ import {
   Spinner,
   Text,
 } from "@chakra-ui/react";
+
 import { Center, RowOrColumn, Column } from "buttered-chakra";
-import React, { useState } from "react";
-import { useTranslation } from "react-i18next";
-import { useRari } from "../../../context/RariContext";
-import { smallUsdFormatter } from "../../../utils/bigUtils";
-import { ClaimRGTModal } from "../../shared/ClaimRGTModal";
-import CopyrightSpacer from "../../shared/CopyrightSpacer";
-import DashboardBox from "../../shared/DashboardBox";
-import ForceAuthModal from "../../shared/ForceAuthModal";
-import { Header } from "../../shared/Header";
 import { NewsAndTwitterLink } from "../MultiPoolPortal";
-import { useQuery } from "react-query";
 import Pool2Modal from "./Pool2Modal";
-import { useIsSmallScreen } from "../../../hooks/useIsSmallScreen";
+import { ClaimRGTModal } from "components/shared/ClaimRGTModal";
+import CopyrightSpacer from "components/shared/CopyrightSpacer";
+import DashboardBox from "components/shared/DashboardBox";
+import ForceAuthModal from "components/shared/ForceAuthModal";
+import { Header } from "components/shared/Header";
+
+
+// Hooks
+import { useTranslation } from "react-i18next";
+import { useRari } from "context/RariContext";
+import { useIsSmallScreen } from "hooks/useIsSmallScreen";
+import { usePool2Balance } from "hooks/pool2/usePool2Balance";
+import { usePool2UnclaimedRGT } from "hooks/pool2/usePool2UnclaimedRGT";
+import { usePool2APR } from "hooks/pool2/usePool2APR";
+import { usePool2TotalStaked } from "hooks/pool2/usePool2TotalStaked";
+import { useHasSushiswapRewardsStarted } from "hooks/pool2/useSushiswapRewards";
+
+// Utils
+import { smallUsdFormatter } from "utils/bigUtils";
+
 
 const Pool2Page = () => {
   const { isAuthed } = useRari();
@@ -169,14 +180,7 @@ export const TotalStaked = () => {
   const { t } = useTranslation();
 
   const { rari } = useRari();
-
-  const { data: totalStaked } = useQuery("pool2TotalStaked", async () => {
-    return parseFloat(
-      rari.web3.utils.fromWei(
-        await rari.governance.rgt.sushiSwapDistributions.totalStakedUsd()
-      )
-    );
-  });
+  const totalStaked = usePool2TotalStaked()
 
   return (
     <Column expand mainAxisAlignment="center" crossAxisAlignment="center">
@@ -197,17 +201,7 @@ export const StartAndEnd = () => {
 
   const { rari } = useRari();
 
-  const { data: hasStarted } = useQuery(
-    "hasSushiswapRewardsStarted",
-    async () => {
-      const block = await rari.web3.eth.getBlockNumber();
-
-      const startingBlock =
-        rari.governance.rgt.sushiSwapDistributions.DISTRIBUTION_START_BLOCK;
-
-      return block >= startingBlock;
-    }
-  );
+  const hasStarted = useHasSushiswapRewardsStarted()
 
   return (
     <Column expand mainAxisAlignment="center" crossAxisAlignment="center">
@@ -246,44 +240,8 @@ const Countdown = ({ endDate }: { endDate: number }) => {
 const YourBalance = () => {
   const { t } = useTranslation();
 
-  const { rari, address } = useRari();
-
-  const { data: balance } = useQuery(address + " pool2Balance", async () => {
-    const SLP = parseFloat(
-      rari.web3.utils.fromWei(
-        await rari.governance.rgt.sushiSwapDistributions.stakingBalanceOf(
-          address
-        )
-      )
-    );
-
-    const {
-      eth: _eth,
-      rgt: _rgt,
-    } = await rari.governance.rgt.sushiSwapDistributions.stakedReservesOf(
-      address
-    );
-
-    return {
-      SLP,
-      hasDeposited: SLP > 0,
-      // @ts-ignore
-      eth: _eth.toString() / 1e18,
-      // @ts-ignore
-      rgt: _rgt.toString() / 1e18,
-    };
-  });
-
-  const { data: earned } = useQuery(
-    address + " pool2Unclaimed RGT",
-    async () => {
-      return parseFloat(
-        rari.web3.utils.fromWei(
-          await rari.governance.rgt.sushiSwapDistributions.getUnclaimed(address)
-        )
-      );
-    }
-  );
+  const balance = usePool2Balance()
+  const earned = usePool2UnclaimedRGT()
 
   const {
     isOpen: isClaimRGTModalOpen,
@@ -342,22 +300,7 @@ const GeneralInfo = () => {
 
   const { rari } = useRari();
 
-  const { data: apr } = useQuery("pool2APR", async () => {
-    const blockNumber = await rari.web3.eth.getBlockNumber();
-    const tvl = await rari.governance.rgt.sushiSwapDistributions.totalStakedUsd();
-
-    return (
-      parseInt(
-        (
-          await rari.governance.rgt.sushiSwapDistributions.getCurrentApr(
-            blockNumber,
-            tvl
-          )
-        ).toString()
-      ) / 1e16
-    ).toFixed(2);
-  });
-
+  const apr = usePool2APR()
   const isMobile = useIsSmallScreen();
 
   const {
